backend/requirements.txt
flask
requests
Pillow
flask-cors
